Improve AES encryption and decryption methods
The AES encryption and decryption methods have been updated to implement PKCS7 padding. The GetSaltBytes method has been modified to take a size parameter for deciding the size of the byte array. Also, the encryption and decryption process now ensure that distinct values are provided for the Key and IV properties. These changes should enhance the robustness of encryption and decryption operations.